feature: check with external booking service if user can be deleted

+++ auth-service/application/user_service.go
@@ -3,6 +3,9 @@ package application
 import (
 	"encoding/json"
 	"errors"
+	booking "github.com/ZMS-DevOps/booking-service/proto"
+	"github.com/mmmajder/zms-devops-auth-service/application/external"
+	"github.com/mmmajder/zms-devops-auth-service/domain"
 	"github.com/mmmajder/zms-devops-auth-service/infrastructure/dto"
 	"log"
 	"net/http"
@@ -13,15 +16,17 @@ type UserService struct {
 	AuthService          *AuthService
 	KeycloakService      *KeycloakService
 	IdentityProviderHost string
+	bookingClient        booking.BookingServiceClient
 }
 
-func NewUserService(httpClient *http.Client, authService *AuthService, keycloakService *KeycloakService, identityProviderHost string) *UserService {
+func NewUserService(httpClient *http.Client, authService *AuthService, keycloakService *KeycloakService, identityProviderHost string, bookingClient booking.BookingServiceClient) *UserService {
 
 	return &UserService{
 		HttpClient:           httpClient,
 		AuthService:          authService,
 		KeycloakService:      keycloakService,
 		IdentityProviderHost: identityProviderHost,
+		bookingClient:        bookingClient,
 	}
 }
 
@@ -45,6 +50,7 @@ func (service *UserService) GetUserById(authorizationHeader, id string) (*dto.Us
 	if err != nil {
 		return nil, err
 	}
+	log.Printf("step 2: %s", id)
 
 	keycloakDTO := &dto.GettingKeycloakUserDTO{}
 	if err := json.NewDecoder(responseBody).Decode(keycloakDTO); err != nil {
@@ -69,9 +75,26 @@ func (service *UserService) UpdateUser(authorizationHeader, id, firstName, lastN
 	return nil
 }
 
-func (service *UserService) DeleteUser(authorizationHeader, id string) error {
-	if err := service.KeycloakService.DeleteKeycloakUser(authorizationHeader, id); err != nil {
-		return err
+func (service *UserService) DeleteUser(authorizationHeader, id, group string) error {
+	var canDeleteUser bool
+	if group == domain.HostRole {
+		response, err := external.IfHostCanBeDeleted(service.bookingClient, id)
+		if err != nil {
+			return err
+		}
+		canDeleteUser = response.Success
+	} else {
+		response, err := external.IfGuestCanBeDeleted(service.bookingClient, id)
+		if err != nil {
+			return err
+		}
+		canDeleteUser = response.Success
+	}
+
+	if canDeleteUser {
+		if err := service.KeycloakService.DeleteKeycloakUser(authorizationHeader, id); err != nil {
+			return err
+		}
 	}
 
 	return nil
